% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrito-electoral.R
\name{assign_seats}
\alias{assign_seats}
\title{Assign seats}
\usage{
assign_seats(x, nseats, method, thr = 0)
}
\arguments{
\item{x}{A matrix or data.frame. Votes for each partie are in columns, rows are
jurisdictions.}

\item{nseats}{A vector with the number of seats per jurisdiction.}

\item{method}{The method used for seats allocation. Available methods are
"dhondt", "webster", "danish", "imperiali", "hill-huntington", "dean",
"mod-saint-lague", "equal-proportions", "adams", "hare", "droop",
"hangenbach-bischoff", "imperial", "mod-imperial", "quotas-remainders"}
}
\value{
A matrix the same dimension of 'x' with the seats allocated to each partie
to every jurisdiction.
}
\description{
Assign seats
}
