% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrito-electoral.R
\name{split_seats}
\alias{split_seats}
\title{Split seats between jurisdictions}
\usage{
split_seats(voters, nseats, fixed = 0, method = "hare")
}
\arguments{
\item{voters}{A vector with the number of voters per jurisdiction.}

\item{nseats}{The total number of seats to split.}

\item{fixed}{The fixed number of seats for each jurisdiction. By default, zero.}

\item{method}{The method used to split seats, by default, "Hare".}
}
\value{
A vector with the number of seats for each jurisdiction.
}
\description{
Split seats between jurisdictions
}
